# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Doorstep Merchant Builder
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthCredentials {
  email: string;
  password: string;
}

export interface MerchantInfo {
  id: number;
  email: string;
}

export interface AuthResponse {
  success: boolean;
  token: string;
  merchant: MerchantInfo;
}

export interface Store {
  id: number;
  merchantId: number;
  name: string;
  logoUrl?: string | null;
  description?: string | null;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateStoreInput {
  name: string;
  logoUrl?: string | null;
  description?: string | null;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
}

export interface Product {
  id: number;
  merchantId: number;
  storeId: number;
  name: string;
  description?: string | null;
  price: number;
  stock: number;
  category?: string | null;
  sku: string;
  brand?: string | null;
  imageUrl?: string | null;
  discountPrice?: number | null;
  weight?: number | null;
  tags?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProductInput {
  name: string;
  description?: string | null;
  price: number;
  stock: number;
  category?: string | null;
  sku: string;
  brand?: string | null;
  imageUrl?: string | null;
  discountPrice?: number | null;
  weight?: number | null;
  tags?: string | null;
}

export interface UpdateProductInput {
  name?: string;
  description?: string | null;
  price?: number;
  stock?: number;
  category?: string | null;
  sku?: string;
  brand?: string | null;
  imageUrl?: string | null;
  discountPrice?: number | null;
  weight?: number | null;
  tags?: string | null;
}

export interface PaginatedProducts {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export type UploadPreviewResponsePreviewRowsItem = { [key: string]: unknown };

export interface RowValidationError {
  row: number;
  sku?: string | null;
  name?: string | null;
  message: string;
}

export interface UploadPreviewResponse {
  success: boolean;
  filename: string;
  totalRows: number;
  previewRows: UploadPreviewResponsePreviewRowsItem[];
  errors: RowValidationError[];
}

export interface UploadLog {
  id: number;
  merchantId: number;
  storeId: number;
  filename: string;
  status: string;
  totalRows: number;
  importedCount: number;
  failedCount: number;
  errors?: string | null;
  createdAt: string;
}

export type Logout200 = {
  success: boolean;
};

export type ListProductsParams = {
search?: string;
category?: string;
sortBy?: ListProductsSortBy;
sortOrder?: ListProductsSortOrder;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListProductsSortBy = typeof ListProductsSortBy[keyof typeof ListProductsSortBy];


export const ListProductsSortBy = {
  name: 'name',
  price: 'price',
  stock: 'stock',
  createdAt: 'createdAt',
} as const;

export type ListProductsSortOrder = typeof ListProductsSortOrder[keyof typeof ListProductsSortOrder];


export const ListProductsSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type BulkDeleteProductsBody = {
  ids: number[];
};

export type BulkDeleteProducts200 = {
  success: boolean;
  count: number;
};

export type DeleteProduct200 = {
  success: boolean;
};

export type UploadImportFileBody = {
  file: Blob;
};

export type ConfirmImportBodyProductsItem = { [key: string]: unknown };

export type ConfirmImportBody = {
  filename: string;
  products: ConfirmImportBodyProductsItem[];
};

export type ConfirmImport200 = {
  success: boolean;
  count: number;
};