# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Doorstep Merchant Builder
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Merchant Signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: Signup successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Merchant Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current merchant profile
      responses:
        "200":
          description: Current merchant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantInfo"
        "410":
          description: Unauthorized
...
  /import/upload:
    post:
      operationId: uploadImportFile
      tags: [import]
      summary: Upload product catalog file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "200":
          description: File parsed successfully, returns row-level preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadPreviewResponse"
  /import/confirm:
    post:
      operationId: confirmImport
      tags: [import]
      summary: Confirm and save imported products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename:
                  type: string
                products:
                  type: array
                  items:
                    type: object
              required: [filename, products]
      responses:
        "200":
          description: Import processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                required: [success, count]
...
components:
  schemas:
...
    AuthCredentials:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
        merchant:
          $ref: "#/components/schemas/MerchantInfo"
      required:
        - success
        - token
        - merchant
    MerchantInfo:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
      required:
        - id
        - email
...
    UploadPreviewResponse:
      type: object
      properties:
        success:
          type: boolean
        filename:
          type: string
        totalRows:
          type: integer
        previewRows:
          type: array
          items:
            type: object
        errors:
          type: array
          items:
            $ref: "#/components/schemas/RowValidationError"
      required:
        - success
        - filename
        - totalRows
        - previewRows
        - errors
    RowValidationError:
      type: object
      properties:
        row:
          type: integer
        sku:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        message:
          type: string
      required:
        - row
        - message
...