# lib/api-zod/src/generated/types/merchantInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Doorstep Merchant Builder
 * OpenAPI spec version: 0.1.0
 */

export interface MerchantInfo {
  id: number;
  email: string;
}