# lib/api-zod/src/generated/types/rowValidationError.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Doorstep Merchant Builder
 * OpenAPI spec version: 0.1.0
 */

export interface RowValidationError {
  row: number;
  sku?: string | null;
  name?: string | null;
  message: string;
}